Change check_binary_integrity() to actually modify the libraries used on disk using the new xor_path_byte() function.
Note that this will only work if the library is being used in
a privileged process with write access to the library path.

Doesn't cope with symlink races.

If the library modification fails in any way the library
forces an abort() core dump giving an error message
listing the library path being modified so the tester can
restore system library integrity.

Signed-off-by: Jeremy Allison <jallison@ciq.com>